Updates   * Cleaned up the form compoents   * Normalized how the components use the StyleInjetor     * Fixes a number of styling issues     * Moved sub-components to new internal folder  * Updated Image component to work same as Form components
