+++ repos/keg-components/configs/aliases.json
@@ -4,7 +4,6 @@
   "KegFilePicker": "src/components/filePicker/filePicker${platform}.js",
   "KegForm": "src/components/form/form/form.js",
   "KegImg": "src/components/image/image.js",
-  "KegInternalImage": "src/components/image/kegImage${platform}.js",
   "KegIcon": "src/components/icon/icon.js",
   "KegIcons": "src/assets/icons",
   "KegIndicator": "src/components/indicator/indicator.js",
