+++ repos/keg-components/src/components/filePicker/filePicker.js
@@ -7,14 +7,16 @@ import { View } from 'KegView'
 import { get } from '@keg-hub/jsutils'
 import { useThemeTypeAsClass } from 'KegTypeAsClass'
 import { StyleInjector } from '@keg-hub/re-theme/styleInjector'
+import { Input as KegInput } from '../internal/input.web'
 
 /**
- * @summary A simple wrapper around the a web input element
- * <br/> This is required because we can't just pass input to the StyleInjector
- * <br/> It requires a React component, and input gets viewed as a variable
- * @param {Object} props - props object. Accepts all standard <input /> props which will be passed to the input element. Additional props:
+ * Wrap the internal component with the Styles Injector Hoc
+ * <br/>This allows us to add the styles as css classes
  */
-const InputComp = props => (<input {...props} />)
+const Input = StyleInjector(
+  KegInput,
+  { displayName: 'FilePickerInput', className: 'keg-file-picker-input' }
+)
 
 /**
  * A component for selecting files from the user's system. Looks better than a basic input element,
@@ -78,11 +80,6 @@ export const FilePicker = React.forwardRef((props, ref) => {
     openOnMount && clickInput()
   }, [])
 
-  const InjectedComp = StyleInjector(
-    InputComp,
-    { displayName: 'FilePicker', className: 'keg-file-picker' }
-  )
-
   return (
     <View
       className={useThemeTypeAsClass(
@@ -109,7 +106,7 @@ export const FilePicker = React.forwardRef((props, ref) => {
       }
 
       { /* this input is hidden from the user, but is still used for selecting a file */ }
-      <InjectedComp
+      <Input
         {...args}
         ref={input => {
           ref && (ref.current = input)
