+++ repos/keg-components/src/components/internal/input.web.js
@@ -1,0 +1,18 @@
+import React from 'react'
+import { StyleInjector } from '@keg-hub/re-theme/styleInjector'
+import { useClassName } from 'KegClassName'
+
+/**
+ * @summary Wrapper around the a web input element
+ * <br/> This is required because we can't just pass input to the StyleInjector
+ * <br/> It requires a React component, and input gets viewed as a variable
+ * @param {Object} props - props object. Accepts all standard <input /> props which will be passed to the input element. Additional props:
+ */
+export const Input = React.forwardRef(({ className, ...props}, ref) => {
+  const classRef = useClassName('keg-input', className, ref)
+
+  return (
+    <input ref={classRef} {...props} />
+  )
+
+})
